+++ CHANGELOG.md
@@ -1,3 +1,18 @@
+## 1.1.0
+
+- Remove `abstract` from class definition to allow inline streamable creation
+
+**Example**:
+
+```dart
+final counter = CachedStreamable<int>(0);
+counter.value++;
+```
+
+### BREAKING CHANGE
+
+- Use `value` instead of `cache`
+
 ## 1.0.2+2
 
 - Update description